% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt_freq.R
\name{cmpt_freq}
\alias{cmpt_freq}
\title{Bayesian Gaussian Process regression with Stan: computation of frequency by MCMC}
\usage{
cmpt_freq(data, rho, sigma, delta = 1/25, M = as.integer(50),
  stan_seed = as.integer(1234), stan_chains = as.integer(3),
  stan_warmup = as.integer(300), stan_thin = as.integer(1),
  stan_iter = as.integer(1300), stan_max_treedepth = as.integer(10))
}
\arguments{
\item{data}{a vector of one dimentional data.}

\item{rho}{numeric >0; how fast the correlation decrease.
(See \strong{Description})}

\item{sigma}{numeric >0; amplitude of frequency plot.
(See \strong{Description})}

\item{delta}{numeric beetween 0 and 1 (defaults to 1/25); the size of
no-data area in frequency plot.
 In order to stabilize computation, both ends of the graph have no data
points. The size is expressed by \code{delta} times size of plot area
 in the x-axis direction.}

\item{M}{integer; degree of discretization computation. x axis of
frequency plot is divided into M grids.
 Defaults to \code{M = 50}, and larger M give accurate frequency plot,
  but the computation takes much time.}

\item{stan_seed}{see \code{?stan} in \code{library(rstan)}}

\item{stan_chains}{see \code{?stan} in \code{library(rstan)}}

\item{stan_warmup}{see \code{?stan} in \code{library(rstan)}}

\item{stan_thin}{see \code{?stan} in \code{library(rstan)}}

\item{stan_iter}{see \code{?stan} in \code{library(rstan)}}

\item{stan_max_treedepth}{see \code{?stan} in \code{library(rstan)}}
}
\value{
list data having follow 5 components:
\itemize{
 \item{\code{$WAIC}: See \strong{Description}}
 \item{\code{$p}: data.frame of the MCMC result. \code{$age_X} is x
 coodinates and \code{$p_mean} is estimate. \code{$p_0025} means that
 the probability that ture frequency is smaller than \code{$p_0025}
 is 0.025. \code{$p_025}, \code{$p_075} and \code{$p_0975} is defined
 in the same way as \code{$p_0025}. \code{$p_n_eff} and \code{$p_Rhat} is
 index to check convergence. In this function, if \code{$p_0025 < ##########} and \code{$p_Rhat < ########}, then consider the MCMC to converge}
 \item{\code{$data}: same as the input data, used in \code{\link{freq_graph}}}
 \item{\code{$rho}: same as the input rho}
 \item{\code{$sigma}: same as the input sigma}
}
}
\description{
The frequency is computed by Gaussian Process (GP) regression. In this
computation, only one-dimensional data can be analyzed, but accurate
analysis is possible. The frequency is expressed by parameter 'p' distrubution,
and p is defined by p = softmax(f), f ~ multi_normal(0, K).
K is computed by Gaussian Kernel; K(x, x') = sigma^2 * exp((x - x')^2 / rho^2).
For explain briefly, sigma and rho govern the smoothness properities.
Sigma is related to amplitude of frequency plot, and rho is how fast
the correlation decrease (it means that if rho become smaller, the possibility is
increasing for having a completely different frequency in distant two points of x).
sigma and rho should be determined by the user, however, there is no clue in
 many cases. Then, function \code{cmpt_freq()} give us a hint; WAIC.
WAIC is one of an estimator of the relative quality of statistical models
for a given set of data. If WAIC is smaller, it can be said that it is a
better model.
}
