% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Osayama.R
\docType{data}
\name{Osayama}
\alias{Osayama}
\title{Detrital zircons in Osyama}
\format{A vector with 45 age data}
\usage{
Osayama
}
\description{
A dataset containing the U-Pb age of detrital zircons of Osayama
}
\keyword{datasets}
