% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tetori.R
\docType{data}
\name{Tetori}
\alias{Tetori}
\title{Tetrital zircons in Akaiwa formation in Tetori group}
\format{A vector with 93 age data}
\usage{
Tetori
}
\description{
A dataset containing the U-Pb age of detrital zircons of Akaiwa formaiton
}
\keyword{datasets}
